using "value" instead of "property"
DSP refers to properties when it is talking about the values of properties, and then talks about the values as 'literals' and 'non-literals'. There is some confusion between "statement" "property" and "value", probably because of the DCAM terminology. If "statements" become "properties" in the RDF sense, then this is simplified.